[PATCH] lspcmcia symlink target is wrong if building with DESTDIR
The lspcmcia symlink is created to $(DESTDIR)$(sbindir)/$(PCCARDCTL); if
you're building with DESTDIR set to a temporary directory in order to
build a binary package for a distribution (which is basically what
DESTDIR is meant for), the symlink in the package ends up pointing into
your build tree. Since lspcmcia and pccardctl are installed to the same
directory anyway, this patch just makes the symlink be relative.

Signed-off-by: Dominik Brodowski <linux@dominikbrodowski.net>